+++ NEWS.md
@@ -1,3 +1,4 @@
 # dsinfo (dev)
 
 * Added a `set_source_script()`
+* You can now also supply an email address to `dsi_sources_from_paths()`